TAKE OUT store configure from main app index.js
