+++ src/index.js
@@ -1,47 +1,21 @@
 import React from "react";
 import ReactDOM from "react-dom";
-
-import {applyMiddleware, createStore, combineReducers} from "redux";
 import {Provider} from "react-redux";
-
-import thunk from "redux-thunk";
-import {compose} from "recompose";
-
-import configureAPI from "./server/configure-API.js";
-
-import {appReducer} from "./reducers/app/index.js";
-import {filmsReducer} from "./reducers/films/index.js";
-import {genreFilterReducer} from "./reducers/genre-filter";
-import {movieListReducer} from "./reducers/movie-list/index.js";
+import {BrowserRouter} from "react-router-dom";
+import configuredStore from "./reducers/configure-store.js";
 
 import App from "./components/app/app.jsx";
+import ScrollToTop from "./components/scroll-to-top/scroll-to-top.jsx";
 import {appOperations} from "./reducers/app/index.js";
-import {BrowserRouter} from "react-router-dom";
 
 
-const init = () => {
-  const api = configureAPI((...args) => store.dispatch(...args));
-
-  const rootReducer = combineReducers({
-    app: appReducer,
-    films: filmsReducer,
-    genreFilter: genreFilterReducer,
-    movieList: movieListReducer
-  });
-
-  const store = createStore(
-      rootReducer,
-      compose(
-          applyMiddleware(thunk.withExtraArgument(api)),
-          window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : (f) => f
-      )
-  );
-
+const init = (store) => {
   store.dispatch(appOperations.setupApp());
 
   ReactDOM.render(
       <Provider store={store}>
         <BrowserRouter>
+          <ScrollToTop/>
           <App/>
         </BrowserRouter>
       </Provider>,
@@ -49,5 +23,5 @@ const init = () => {
   );
 };
 
-init();
+init(configuredStore);
